Dockerfile eslint test, docker-publish.yml fix, circular dependency fix
